fixed EratosthenesSieve

+++ src/main/de/psb/sieve/EratosthenesSieve.java
@@ -10,8 +10,8 @@ public class EratosthenesSieve implements IPrimeSieve {
         boolean[] sieve = new boolean[limit + 1];
         sieve[0] = sieve[1] = true;
 
-        for (int i = 0; i * i <= limit; i++)
-            if (!sieve[i])
+        for (long i = 0; i * i <= limit; i++)
+            if (!sieve[(int) i])
                 for (long j = i * 2; j <= limit; j += i)
                     sieve[(int) j] = true;
